// web/src/api/generated/model/countryDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { CalendarKind } from './calendarKind';
import type { CapabilitiesDto } from './capabilitiesDto';
import type { ChartTemplateDto } from './chartTemplateDto';
import type { CurrencyUnitsDto } from './currencyUnitsDto';
import type { DayOfWeek } from './dayOfWeek';
import type { TaxRegistrationRuleDto } from './taxRegistrationRuleDto';

export interface CountryDto {
  code: string;
  nameEn: string;
  nameAr: string;
  languages: string[];
  currency: CurrencyUnitsDto;
  calendars: CalendarKind[];
  defaultWeekend: DayOfWeek[];
  defaultFiscalYearStartMonth: number;
  taxRegistration: TaxRegistrationRuleDto[];
  chartsOfAccounts: ChartTemplateDto[];
  bilingualPrintRequired: boolean;
  capabilities: CapabilitiesDto;
}